Create Docker image to run migrations in production
Add Dockerfile based on the boxfuse/flyway Docker image, which can
be used to run migrations in ECS against the deployed database.

This will make it fairly straightforward to run migrations in Alpha:
a developer can go to ECS in the AWS Console and start a new task
to run the migrations.

In Beta, we should revisit this approach and find a way to run
migrations through the CI server.