Switch File IDs from ints to UUIDs
Use UUIDs rather than integers to identify files in the DB and in GraphQL
requests.

This will make it possible for the frontend to generate a UUID file ID and
then upload each file to S3 with a UUID name. This means we won't expose
sequential file IDs to the user, which might look confusing if they don't
see sequential IDs for their own files.