Manually define the GraphQL FileType
Replace the automatic type generation with a manualy definition that
specifies all the fields on File.

This will make it easier to make the `fileStatus` property deferred, so
that it is only fetched if GraphQL requires it and so that it can be
added to any query that returns the FileType.