Add progress column to consignments table
Add column to keep track of a consignment's progression through the
transfer project.

This ought to be an enum column to restrict the allowed values, but
it's simpler to implement it as a varchar for the prototype. Adding
it as an enum will require a lot more work in the Slick table
definitions.