Run consignment export task when a transfer is confirmed
When a user confirms that a consignment can be transferred to TNA, trigger
the export task in TNA.

In Beta, this should be decoupled from the API using SNS as an intermediary.
This will remove all the ECS configuration from the API, and will let us add
retry logic if an export fails. But that needs extra infrastructure, so for
now just trigger the task from the API to keep the prototype simple.