Use connection pool to connect to PostgreSQL DB
Create a Slick DB connection using `Database.forConfig` rather than
`Database.forUrl`. Under the surface, this will use
`slick.jdbc.hikaricp.HikariCPJdbcDataSource$` to create a connection
pool, rather than using `DataSourceJdbcDataSource`, which has a
limited number of connections but does not use a pool.

The purpose of this change is to see whether it fixes the
`java.net.ConnectException` which we get when running the export
process for consignments with a large number of files, and which may
be caused by the default data source not managing the connections well.